workaround: the current db instance can only handle string with up to 255 characters for user defined types
the reason for this behaviour is not known to me. The default should be 1M for any LOB type - but that does not seem to apply for methods to instantiate user defined types. Setting them (the limits) manually did not help either.

Undo this commit if the problem was solved!