fix: Db2StatementTester#testExecuteQuery
The test used a schema (TESTDB2) that is not created during test setup. It also referenced some types (TUDTS and TUDTC) that are not defined during setup. This commit fixes the test - but it is not obvious to me what the intenction of this test was in the first place. It just creates a table and then makes a query. is it important to have these specific column/ types? why?